requirements.txt
asyncio==3.4.3
Authlib==0.15.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
h11==0.12.0
httpcore==0.13.0
httpx==0.18.0
idna==2.10
prompt-toolkit==3.0.18
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.17.0
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
tda-api==1.2.0
tpqoa @ git+https://github.com/yhilpisch/tpqoa@7f75b3399ec5b05ad8c543e95c815a76372db98d
ujson==4.0.2
urllib3==1.26.4
v20==3.0.25.0
wcwidth==0.2.5
websockets==8.1
wincertstore==0.2
